Move get_track_repost_intersection_users
